+++ src/game/config/layers.js
@@ -1,9 +1,14 @@
-const layers = [
+const LAYERS = [
     {
         name: 'main',
         size: 'auto',
         zindex: 100
+    },
+    {
+        name: 'static',
+        size: 'auto',
+        zindex: 80
     }
 ];
 
-export { layers };
+export { LAYERS };