+++ src/game/game.js
@@ -1,8 +1,0 @@
-import { layers } from './config/layers';
-
-const scale = 16 / 9;
-
-export {
-    layers,
-    scale
-}